MLK-24283-3 usb: cdns3: gadget: the TD_SIZE entry is 16 bit at TRB
The TD_SIZE entry is 16 bit at TRB for HS/FS, it fixed the issue
if the request length is 16384 bytes for FS transfer, at this case,
the TD_SIZE is 0 at current code.

+++ drivers/usb/cdns3/gadget.c
@@ -613,7 +613,7 @@ int cdns3_ep_run_transfer(struct cdns3_endpoint *priv_ep,
 	u32 control;
 	int pcs;
 	struct scatterlist *s = NULL;
-	u8 td_size;
+	u16 td_size;
 
 	if (priv_ep->type == USB_ENDPOINT_XFER_ISOC)
 		num_trb = priv_ep->interval;
